Fix Quit + Line Drawing For Editor

+++ dev/include/primitives.h
@@ -3,7 +3,7 @@
 
 #include "bool.h"
 
-void drawLine(int x1, int y1, int x2, int y2);
+void drawLine(float x1, float y1, float x2, float y2);
 void drawSquare();
 void drawSquareTexture();
 void drawSquareBorder();
